Validate sensitive environment variable config when creating env vars (#185)
Turns out the API elides non-sensitive values into being sensitive if
the team has a sensitiveEnvironmentVariablePolicy set on it.

To prevent this from causing provider issues when the `sensitive` field
is explicitly set to `false`, warn at plan time that the resource being
created does not match the team policy.

Closes #183

+++ sweep/main.go
@@ -26,6 +26,7 @@ func main() {
 		//lintignore:R009
 		panic("VERCEL_TERRAFORM_TESTING_DOMAIN environment variable not set")
 	}
+	clearDNS := os.Getenv("VERCEL_TERRAFORM_CLEAR_DNS") != ""
 	ctx := context.Background()
 
 	// delete both for the testing team, and for without a team
@@ -34,10 +35,12 @@ func main() {
 		//lintignore:R009
 		panic(err)
 	}
-	err = deleteAllDNSRecords(ctx, c, domain, teamID)
-	if err != nil {
-		//lintignore:R009
-		panic(err)
+	if clearDNS {
+		err = deleteAllDNSRecords(ctx, c, domain, teamID)
+		if err != nil {
+			//lintignore:R009
+			panic(err)
+		}
 	}
 	err = deleteAllSharedEnvironmentVariables(ctx, c, teamID)
 	if err != nil {
